Add loading screen to connector
