+++ packages/yoroi-extension/app/ergo-connector/actions/connector-actions.js
@@ -7,7 +7,7 @@ export default class ConnectorActions {
   getResponse: AsyncAction<void> = new AsyncAction();
   getSigningMsg: AsyncAction<void> = new AsyncAction();
   refreshActiveSites: AsyncAction<void> = new AsyncAction();
-  getWallets: Action<void> = new Action();
+  refreshWallets: AsyncAction<void> = new AsyncAction();
   closeWindow: Action<void> = new Action();
   getConnectorWhitelist: AsyncAction<void> = new AsyncAction();
   updateConnectorWhitelist: AsyncAction<{|
