+++ packages/yoroi-extension/app/ergo-connector/containers/ConnectWebsitesContainer.js
@@ -22,7 +22,7 @@ export default class ConnectWebsitesContainer extends Component<
   InjectedOrGeneratedConnector<GeneratedData>
 > {
   async componentDidMount() {
-    this.generated.actions.connector.getWallets.trigger();
+    this.generated.actions.connector.refreshWallets.trigger();
     this.generated.actions.connector.refreshActiveSites.trigger();
     await this.generated.actions.connector.getConnectorWhitelist.trigger();
   }
@@ -72,8 +72,8 @@ export default class ConnectWebsitesContainer extends Component<
   @computed get generated(): {|
     actions: {|
       connector: {|
-        getWallets: {|
-          trigger: (params: void) => void,
+        refreshWallets: {|
+          trigger: (params: void) => Promise<void>,
         |},
         refreshActiveSites: {|
           trigger: (params: void) => Promise<void>,
@@ -115,7 +115,7 @@ export default class ConnectWebsitesContainer extends Component<
       },
       actions: {
         connector: {
-          getWallets: { trigger: actions.connector.getWallets.trigger },
+          refreshWallets: { trigger: actions.connector.refreshWallets.trigger },
           refreshActiveSites: { trigger: actions.connector.refreshActiveSites.trigger },
           removeWalletFromWhitelist: {
             trigger: actions.connector.removeWalletFromWhitelist.trigger,
