+++ packages/yoroi-extension/app/ergo-connector/containers/ConnectWebsitesContainer.stories.js
@@ -70,7 +70,7 @@ const genBaseProps: {|
     },
     actions: {
       connector: {
-        getWallets: { trigger: action('getWallets') },
+        refreshWallets: { trigger: async (req) => action('refreshWallets')(req) },
         removeWalletFromWhitelist: { trigger: async (req) => action('removeWalletFromWhitelist')(req) },
         refreshActiveSites: { trigger: async (req) => action('refreshActiveSites')(req) },
         getConnectorWhitelist: { trigger: async (req) => action('getConnectorWhitelist')(req) },
