+++ packages/yoroi-extension/app/ergo-connector/stores/index.js
@@ -5,6 +5,7 @@ import type { Api } from '../../api/index';
 import UiNotificationsStore from '../../stores/toplevel/UiNotificationsStore';
 import UiDialogsStore from '../../stores/toplevel/UiDialogsStore';
 import ConnectorStore from './ConnectorStore';
+import ConnectorLoadingStore from './ConnectorLoadingStore';
 import type { ActionsMap } from '../actions';
 
 /** Map of var name to class. Allows dynamic lookup of class so we can init all stores one loop */
@@ -12,6 +13,7 @@ const storeClasses = Object.freeze({
   profile: ProfileStore,
   uiDialogs: UiDialogsStore,
   uiNotifications: UiNotificationsStore,
+  loading: ConnectorLoadingStore,
   connector: ConnectorStore,
 });
 
@@ -19,6 +21,7 @@ export type StoresMap = {|
   profile: ProfileStore,
   uiDialogs: UiDialogsStore<{||}, ActionsMap>,
   uiNotifications: UiNotificationsStore<{||}, ActionsMap>,
+  loading: ConnectorLoadingStore,
   connector: ConnectorStore,
 |};
 
@@ -27,6 +30,7 @@ const stores: WithNullableFields<StoresMap> = observable({
   profile: null,
   uiDialogs: null,
   uiNotifications: null,
+  loading: null,
   connector: null,
 });
 
@@ -47,6 +51,10 @@ export default (action(
     });
 
     const loadedStores: StoresMap = (stores: any);
+
+    // Perform load after all setup is done to ensure migration can modify store state
+    loadedStores.loading.load();
+
     return loadedStores;
   }
 ): (Api, ActionsMap) => StoresMap);
