+++ packages/yoroi-extension/app/ergo-connector/types.js
@@ -1,5 +1,6 @@
 // @flow
 
+// TODO: delete this and replace it with a Request object
 export const LoadingWalletStates = Object.freeze({
   IDLE: 0,
   PENDING: 1,
